+++ src/main/java/bobthebot/bob/BobTheBot.java
@@ -1,25 +1,35 @@
-package bob;
+package bobthebot.bob;
 
-import utils.Storage;
-import utils.Parser;
-import utils.Ui;
-import tasks.ToDoList;
-import exceptions.BobException;
+import bobthebot.exceptions.BobException;
+import bobthebot.utils.Storage;
+import bobthebot.utils.Parser;
+import bobthebot.utils.Ui;
+import bobthebot.tasks.ToDoList;
 
 import java.io.IOException;
 import java.util.Scanner;
 
+/**
+ * BobTheBot is a friendly CLI Bot that can help with managing tasks.
+ * */
 public class BobTheBot {
     private Storage storage;
     private ToDoList list;
     private Parser parser;
 
+    /**
+     * Constructor for instance of BobTheBot
+     * @param filePath of data.txt file where tasks are stored.
+     * */
     public BobTheBot(String filePath) throws IOException {
         this.storage = new Storage(filePath);
         this.list = new ToDoList(this.storage.load(), storage);
         this.parser = new Parser();
     }
 
+    /**
+     * Command to run BobTheBot.
+     * */
     public void run() throws BobException {
         Ui.welcome();
         Scanner scanner = new Scanner(System.in);
@@ -35,7 +45,10 @@ public class BobTheBot {
         scanner.close();
     }
 
-    public static void main(String[] args) throws BobException, IOException {
+    /**
+     * Main class where constructor is called and BobTheBot is run.
+     * */
+    public static void main(String[] args) throws IOException, BobException {
         new BobTheBot("./../../data/data.txt").run();
     }
 }
