+++ src/main/java/bobthebot/exceptions/BobException.java
@@ -1,5 +1,8 @@
-package exceptions;
+package bobthebot.exceptions;
 
+/**
+ * Custom exception class for handling exceptions specific to BobTheBot
+ * */
 public class BobException extends Exception {
     public BobException(String errorMessage) {
         super(errorMessage);
