+++ src/main/java/bobthebot/tasks/ToDoList.java
@@ -1,29 +1,34 @@
-package tasks;
+package bobthebot.tasks;
 
-import utils.Storage;
-import tasks.Task;
-import exceptions.BobException;
-import utils.Ui;
-import tasks.Todo;
-import tasks.Event;
-import tasks.Deadline;
+import bobthebot.utils.Storage;
+import bobthebot.utils.Ui;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 
+/**
+ * Class which represents a ToDoList.
+ * */
 public class ToDoList {
     private ArrayList<Task> list;
     private Storage storage;
 
-    /* Define constructor for to do list */
+    /**
+     * Constructs instance of ToDoList.
+     * @param list of tasks. List of tasks can be empty or contains elements taken from storage.
+     * @param storage Storage from which
+     * */
     public ToDoList(ArrayList<Task> list, Storage storage) {
         this.list = list;
         this.storage = storage;
     }
 
-    /* Method for adding items to the list */
+    /**
+     * Method for adding items to the list. Method also handles the logic for the type of task to be added.
+     * @param command A String which contains the command about the task to be added.
+     * */
     public void addTask(String command) {
         if (command.startsWith("todo")) {
             command = command.replace("todo", "").trim();
@@ -156,6 +161,10 @@ public class ToDoList {
         }
     }
 
+    /**
+     * Method for deleting a specific event from the list, and updating the storage.
+     * @param index Specifies 0 index of task to be deleted.
+     * */
     /* Method for deleting a specific event */
     public void deleteTask(int index) {
         Ui.taskDeletedMessage(index, this);
@@ -163,40 +172,47 @@ public class ToDoList {
         storage.store(list);
     }
 
-    /*
-     * Method to mark a certain item in the list as done
-     * takes in the 1 index of the task
-     */
+    /**
+     * Method for marking a specific event from the list as done, and updating the storage.
+     * @param index Specifies 1 index of task to be marked as done.
+     * */
     public void markItemDone(int index) {
         this.list.get(index - 1).markDone();
         Ui.markItemDoneMessage(this, index - 1); // takes in 0 index
         storage.store(list);
     }
 
-    /*
-     * Method to mark a certain item in the list as undone
-     * takes in the 1 index of the task
-     */
+    /**
+     * Method for marking a specific event from the list as undone, and updating the storage.
+     * @param index Specifies 1 index of task to be marked as undone.
+     * */
     public void markItemUndone(int index) {
         this.list.get(index - 1).markUndone();
         Ui.markItemUndoneMessage(this, index - 1); // takes in 0 index
         storage.store(list);
     }
 
-    /* Method that gets the length of the list */
+    /**
+     * Method that returns the number of items in the ToDo List.
+     * @return An int representing the number of items in the ToDo List.
+     * */
     public int getLength() {
         return this.list.size();
     }
 
-    /*
-     * Method to get a certain task in the list
-     * Takes in the 0 index of the task
-     */
+    /**
+     * Method that returns the specified task.
+     * @param index 0 index of the task specified.
+     * @return Task at the specified index.
+     * */
     public Task getTask(int index) {
         return this.list.get(index);
     }
 
-    /* Method for printing items in the list */
+    /**
+     * Returns a String containing all the elements in the list.
+     * @return String containing all the elements in the list.
+     * */
     @Override
     public String toString() {
         int numOfElements = this.list.size();
